Remove export call from tiles usage policy
The export call is restricted to openstreetmap.org usage only via
a totp token. As it is no longer available for general usage (and
in practice, hasn't before that due to load reasons), it does not
belong on the tile usage policy.

Additionally, it is served from a different domain name, so the
fact that it shares some resources with the tiles is an implementation
detail behind the two services.