Use kramdown to apply bootstrap classes to tables
Bootstrap doesn't apply any styling to tables by default, which can
make them hard to read. But kramdown allows us to add classes to
elements in markdown, without having to convert anything to html.

This applies the table and table-hover classes to the financial tables.
It also backports the w-auto class from later versions of bootstrap
to avoid making the tables super-wide and therefore harder to read.

I've also added spaces between the coloured icons and the text, again
for legibility.