Add the OWG travel policy, as adopted 2023-03-09
In the https://osmfoundation.org/wiki/Operations/Minutes/2023-03-09#Travel_policy
meeting we did a live edit of the travel policy and agreed on it.
Reviewing the minutes, I see that the changes suggested were
some copyediting, making the purpose of the travel more general
and not strictly related to hardware, and adding vehicle rental
as an option for travel (e.g. for moving servers).

Looking at the text from the hackpad, all of these changes are
reflected in the policy.

It's not clear if we formally adopted it as a policy, but it was
agreed on at the time and may have fallen through the cracks in
the minutes.